requirements.txt
Django
django-environ
djangorestframework
django-phonenumber-field
phonenumbers
